% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deploy.R
\name{deploy}
\alias{deploy}
\title{deploy}
\usage{
deploy(
  to_execute = "cron",
  model_funs = c(clmodels::BAV, clmodels::GAN, clmodels::VIM, clmodels::CAN),
  model_units = c(1, 1, 1, 1.5),
  max_units = 2.5,
  wealth_scale = 1,
  full_path_to_td_credientials = "~/td"
)
}
\arguments{
\item{to_execute}{character, either 'cron', 'force', or 'report', Default: 'cron'}

\item{model_funs}{vector, the unquoted function calls (without parenthesis) for each model}

\item{model_units}{numeric, a numeric of length equal to model_funs, indicating the number of units to allocate to each model, where the first value entered corresponds to the first (leftmost) model function call, and so on until either series is complete}

\item{max_units}{numeric, the maximum number of units which may be allocated to the portfolio, Default: 2.5}

\item{wealth_scale}{numeric, from 0-1, what percentage of wealth should be allocated to the portfolio, Default: 1.0}

\item{full_path_to_td_credientials}{character, the full path to the directory containg the files essential for API access (see References below), Default: '~/td'}
}
\value{
Nothing, unless to_execute = 'report', in which case output is a list of data tables, a matrix, and a numeric which indicate, respectively, assets to sell/buy, VIX EMA value, and the BAV SMA value.
}
\description{
Auto-trade with TD's API.
}
\details{
Run this every day at 15:59 using a cronjob.  Use in conjuction with with the updateprices and clmodels packages.
}
\examples{
\dontrun{
if(interactive()){
 ## just show a report
 deploy('report')
 ## most should set their own defaults to run the below every day
 cldeploy::deploy()
 ## increase units for GAN (relative to defaults)
 deploy(model_units = c(1, 2, 1, 1.5))
 ## stop using GAN and VIM
 deploy(
   model_funs = c(clmodels::BAV, clmodels::CAN),
   model_units = c(1, 1.5))
 ## scale in when this strategy is still fresh
 deploy(wealth_scale = 0.5)
 ## the path can end with a slash, or not
 deploy(full_path_to_td_credientials = '~/td/')
 }
}
}
\references{
\url{https://github.com/exploringfinance/rameritrade/}
}
